+++ src/processor/object-parsers/default-object-parsers.ts
@@ -1,4 +1,3 @@
 import type { ObjectParser } from ".";
-import { numberWithUnitParser } from "./parsers/number-with-unit";
 
-export const defaultObjectParsers: ObjectParser[] = [numberWithUnitParser];
+export const defaultObjectParsers: ObjectParser[] = [];